Move some helpers over to a utils file.
